Refactored relay, fixed a bug and documented adding new devices (#64)
* refactor: move relay installation code to its own file

* fix: pull images on relay to ensure that they're the latest ones

* doc: documented the process of adding support for a new device

* fix: add missing import to relay
